+++ src/com/vendsy/bartsy/utils/Constants.java
@@ -16,13 +16,13 @@ public class Constants {
 	public static final String ASSETS_PATH = "file:///android_asset/";
 	
 	// For now use hard-coded value of tax
-	public static final float taxRate = (float) (9.5) / (float) (100.0);
+	public static final double taxRate = (double) (9.5) / (double) (100.0);
 	
 	public static final long monitorFrequency = 60000 ; // frequency in which to run the background service, in ms
 	public static final int timoutDelay = 2; // how many more minutes to delay a local timeout from the server timeout
 
 	public static final boolean bundleOrders = false;
 	
-	public static final float defaultTip = (float) 20 / (float) 100;
+	public static final double defaultTip = (double) 20 / (double) 100;
 
 }
